document.go only focuses on macros, not variables
I've come to the conclusion that document.go was way to big and was trying to handle way to much. As long as I have a decent Read() function in document.go, I can pipe it into another bit of logic that can handle the interpretation and expansion of variables. I will thow some documentation together next commit.