I'm passing a test now, but we have a reopen bug
In a previous effort to reduce excesive amount of file opens, I made it so if the same file was open twice (in this case, tests/documents/test.txt), it would just re-use the previous file descriptor.
However, when streaming the document that includes that file twice the problem is that the first read through will cause the curos to be at the EOF, and on the second go-around it does not read anything.