lots of venture programming.
I was mostly getting variables being convered into defeinitions while streaming taken care of.

 I've noticed some architectual issues (see comments)

 I've also added more descriptive commments in the confusiing parts.

 I cleaned up doccomp.go too... function before feature.

Added the 'nonConvertedFile' struct that is basically a file that does not need converting
 I got done with nonConvertedFile.Read but now I just need to write doc.define to get proc vars and normal vars. After that we're ready for more testing.

Removed ProcessLoader... don't think I'll need it for a while.

Made GetVariables return a map instead of an array... not sure if that was a good idae (slower)

I made the availbe processors package-wide but I also made it so documents have their own scope via 'attachProcessor'. Not sure if that is a good idea either.