Okay. We're looking pretty good for the c connector.
All that's left is to go thorugh the rest of vorlage and apply the updates that I've made to processor.go

Also, as of of this commit, what I do is run the following:

 1. cd c.src/ && gcc -shared -o libtest.so -fpic testproc.c; cd ..
 2. go test procload-c_test.go procload-c.go processors.go procload-c-exports.go logging.go

Running those two commands should not produce any build errors and should output all the evidence needed to prove that the c linker is working well.
Also on the 'go test' function I made it fail deliberately so I can see the log output.