BROKEN: so I need to make a commit about this...
So when I was structuring the vorlage processor interface... the structure of "RequestInfo" had taken a different definition. It used to be a struct that dictated the information about the entire request. Now, the "RequestInfo" struct holds information that is only particular to a given processor. So the next few commits will be reflecting this change.