added some implementation-defined inputs to be present in every page request.
I reflected the use of php's 'superglobals'. These inputs are not documented anywhere... I'm just putting them here because dossibay and arbeitgeber are both in need of them at this moment.

in the future of vorlage. implementations hould have the option to pass in custom arguments to the processors that is not through the user.

Even though these inputs are technically user input... they lay in an awkward middle ground where the user supplies them but doesn't have the ability to lie about them.